# swagger.yml
openapi: 3.0.0
info:
  title: "RP Flask REST API"
  description: "An API about todo Application"
  version: "1.0.0"
servers:
  - url: "/api"
components:
  schemas:
...
    EditTaskData:
      type: object
      properties:
        old_title:
          type: string
        new_title:
          type: string
        old_task_name:
          type: string
        new_task_name:
          type: string
      required:
        - old_title
        - new_title
        - old_task_name
        - new_task_name
    TaskData:
      type: object
      properties:
        username:
          type: string
        title:
          type: string
        task_name:
          type: string
      required:
        - username
        - title
        - task_name
...
  securitySchemes:
    jwt_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: jwt_manager.security_handler
paths:
...
  /user/task/create:
    post:
      operationId: tasks.create_task
      tags:
        - Task
      summary: "Creates a Task"
      requestBody:
        description: "A user Creates a Task"
        required: True
        content:
          application/x-www-form-urlencoded:
            schema:
              x-body-name: "task-data"
              $ref: "#/components/schemas/TaskData"
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: Task Created successfully
...
  /user/task/update:
    put:
      operationId: tasks.update_task
      tags:
        - Task
      summary: "Updates a user Task"
      requestBody:
        description: "A user Updates a Task"
        required: True
        content:
          application/x-www-form-urlencoded:
            schema:
              x-body-name: "taskdata"
              $ref: "#/components/schemas/EditTaskData"
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: Profile Retrieved successfully
  /user/{username}/tasks/{task_name}/get:
    get:
      operationId: tasks.get_task
      tags:
        - Task
      summary: "Retrieves a user task"
      parameters:
        - name: username
          in: path
          description: The username of the user
          required: true
          schema:
            type: string
        - name: task_name
          in: path
          description: The name of the task
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: Task Retrieved successfully
  /user/{username}/tasks/get/all:
    get:
      operationId: tasks.get_all_user_tasks
      tags:
        - Task
      summary: "Retrieves all user's task"
      parameters:
        - name: username
          in: path
          description: The username of the user
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: Task Retrieved successfully
  /user/{username}/tasks/{task_name}/delete:
    delete:
      operationId: tasks.delete_task
      tags:
        - Task
      summary: "Deletes a user's Task"
      parameters:
        - name: username
          in: path
          description: The username of the User
          required: true
          schema:
            type: string
        - name: task_name
          in: path
          description: The task_name of the task to delete
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: Task has been deleted successfully
  /user/{username}/tasks/delete/all:
    delete:
      operationId: tasks.clear_all_user_tasks
      tags:
        - Task
      summary: "Deletes all user's Tasks"
      parameters:
        - name: username
          in: path
          description: The username of the user
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: User Tasks has been deleted successfully
  /user/tasks/{task_name}/assign/{assign_to}:
    put:
      operationId: tasks.assign_task
      tags:
        - Task Assign
      summary: "Assign tasks"
      parameters:
        - name: assign_to
          in: path
          description: Assign task to a user
          required: true
          schema:
            type: string
        - name: task_name
          in: path
          description: task name to assign
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: User Task has been assigned successfully
  /user/tasks/{task_name}/unassign/{unassigned_from}:
    put:
      operationId: tasks.unassign_task
      tags:
        - Task Assign
      summary: "Unassign tasks"
      parameters:
        - name: unassigned_from
          in: path
          description: unassign task from a user
          required: true
          schema:
            type: string
        - name: task_name
          in: path
          description: name of task
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: User Task has been successfully unassigned
...
  /user/{assignee}/tasks/assigned/all:
    get:
      operationId: tasks.show_assigned_tasks
      tags:
        - Task Assign
      summary: "Retrieves a list of assigned tasks"
      parameters:
        - name: assignee
          in: path
          description: The username of the assignee
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: list of assigned tasks retrieved successfully
  /user/tasks/{task_name}assigned/all:
    get:
      operationId: tasks.show_assigned_users
      tags:
        - Task Assign
      summary: "Retrieves a list of assigned users for a task"
      parameters:
        - name: task_name
          in: path
          description: The task_name of the task
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: list of assigned users retrieved successfully
...